pipelines/pyproject.toml
[tool.poetry]
name = "pipelines"
version = "0.1.0"
description = "Pipelines microservice for BadgerDoc project"
authors = ["BadgerDoc team"]

[tool.poetry.dependencies]
python = "^3.8.1"
python-dotenv = "^0.19.0"
fastapi = "^0.70.0"
pydantic = "^1.8.2"
psycopg2-binary = "^2.9.1"
sqlalchemy = "1.3.24"
SQLAlchemy-Utils = "^0.37.8"
uvicorn = {extras = ["standard"], version = "^0.15.0"}
minio = "^7.1.1"
alembic = "^1.7.5"
requests = "^2.26.0"
PyJWT = {extras = ["crypto"], version = "^2.3.0"}
urllib3 = "^1.26.8"
setuptools = "^60.5.0"
aiokafka = "^0.7.2"
apache-airflow-client = "^2.7.0"

[tool.poetry.dev-dependencies]
pre-commit = "^2.14.0"
pytest = "^6.2.4"
pytest-asyncio = "^0.15.1"
freezegun = "^1.1.0"
pytest-cov = "^3.0.0"
isort = "^5.9.3"
black = "^22.10.0"
mypy = "^0.910"
flake8 = "^6.0.0"
types-requests = "^2.28.11.15"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
